dynamic/requirements.txt
matplotlib>=3.4.3
numpy>=1.21.2
torch>=1.9.1
scipy>=1.7.1
neuralpredictors==0.1.0
nnfabrik==0.1.0
tqdm>=4.62.3
Pillow>=8.3.2
setuptools>=58.1.0
scikit-image>=0.18.3
h5py>=3.4.0
gitpython
PyYAML>=6.0
scikit-learn>=1.0
opencv-python>=4.5.5.62
pandas>=1.4.1
rawpy>=0.17.0
seaborn>=0.11.2
torchviz>=0.0.2
torchvision>=0.10.1
datajoint>=0.13.2
moviepy
imageio
wandb
